Update kitty shortcuts for new windows/tabs in the CWD
This patch updates the kitty shortcuts for new windows/tabs in the CWD
to use the CWD of the shell rather than the process. This works around
an annoying bug in nvim where new tabs always open in the pyright dist
directory when editing Python files because that's the CWD of the
current process.

https://github.com/kovidgoyal/kitty/issues/5057